chore(lint): fix file-length + unparam, address review nits
- Extract Helm history/revision loaders into commands_load_helm_history.go
  so commands_load.go drops back under the 800-line limit (revive
  file-length-limit) after the Critical-priority change.
- dashReq test helper drops its always-"c1" kctx param (unparam).
- TestInvalidatePreviewForCursorChange_KeepsHighPriorityView uses a
  non-zero requestGen so the High-priority survival is tested directly,
  not via the Gen==0 exemption (CodeRabbit).
